+++ tests/integration/auth/keeper/msg_server_test.go
@@ -18,10 +18,8 @@ import (
 	"cosmossdk.io/x/auth"
 	authkeeper "cosmossdk.io/x/auth/keeper"
 	authsims "cosmossdk.io/x/auth/simulation"
-	"cosmossdk.io/x/auth/types"
 	authtypes "cosmossdk.io/x/auth/types"
 	"cosmossdk.io/x/bank"
-	"cosmossdk.io/x/bank/keeper"
 	bankkeeper "cosmossdk.io/x/bank/keeper"
 	"cosmossdk.io/x/bank/testutil"
 	banktypes "cosmossdk.io/x/bank/types"
@@ -77,9 +75,7 @@ func initFixture(t *testing.T) *fixture {
 	newCtx := sdk.NewContext(cms, true, logger)
 
 	router := baseapp.NewMsgServiceRouter()
-	router.SetInterfaceRegistry(cdc.InterfaceRegistry())
 	queryRouter := baseapp.NewGRPCQueryRouter()
-	queryRouter.SetInterfaceRegistry(cdc.InterfaceRegistry())
 
 	handler := directHandler{}
 	account := baseaccount.NewAccount("base", signing.NewHandlerMap(handler))
@@ -108,7 +104,7 @@ func initFixture(t *testing.T) *fixture {
 	blockedAddresses := map[string]bool{
 		authKeeper.GetAuthority(): false,
 	}
-	bankKeeper := keeper.NewBaseKeeper(
+	bankKeeper := bankkeeper.NewBaseKeeper(
 		runtime.NewEnvironment(runtime.NewKVStoreService(keys[banktypes.StoreKey]), log.NewNopLogger()),
 		cdc,
 		authKeeper,
@@ -130,7 +126,7 @@ func initFixture(t *testing.T) *fixture {
 			accounts.ModuleName:  accountsModule,
 			authtypes.ModuleName: authModule,
 			banktypes.ModuleName: bankModule,
-		})
+		}, router, queryRouter)
 
 	authtypes.RegisterInterfaces(cdc.InterfaceRegistry())
 	banktypes.RegisterInterfaces(cdc.InterfaceRegistry())
@@ -186,13 +182,13 @@ func TestAsyncExec(t *testing.T) {
 
 	testCases := []struct {
 		name      string
-		req       *types.MsgNonAtomicExec
+		req       *authtypes.MsgNonAtomicExec
 		expectErr bool
 		expErrMsg string
 	}{
 		{
 			name: "empty signer address",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: "",
 				Msgs:   []*codectypes.Any{},
 			},
@@ -201,7 +197,7 @@ func TestAsyncExec(t *testing.T) {
 		},
 		{
 			name: "invalid signer address",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: "invalid",
 				Msgs:   []*codectypes.Any{},
 			},
@@ -210,7 +206,7 @@ func TestAsyncExec(t *testing.T) {
 		},
 		{
 			name: "empty msgs",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: addrs[0].String(),
 				Msgs:   []*codectypes.Any{},
 			},
@@ -219,7 +215,7 @@ func TestAsyncExec(t *testing.T) {
 		},
 		{
 			name: "valid msg",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: addrs[0].String(),
 				Msgs:   []*codectypes.Any{msgAny},
 			},
@@ -227,7 +223,7 @@ func TestAsyncExec(t *testing.T) {
 		},
 		{
 			name: "multiple messages being executed",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: addrs[0].String(),
 				Msgs:   []*codectypes.Any{msgAny, msgAny},
 			},
@@ -235,7 +231,7 @@ func TestAsyncExec(t *testing.T) {
 		},
 		{
 			name: "multiple messages with different signers",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: addrs[0].String(),
 				Msgs:   []*codectypes.Any{msgAny, msgAny2},
 			},
@@ -244,7 +240,7 @@ func TestAsyncExec(t *testing.T) {
 		},
 		{
 			name: "multi msg with one failing being executed",
-			req: &types.MsgNonAtomicExec{
+			req: &authtypes.MsgNonAtomicExec{
 				Signer: addrs[0].String(),
 				Msgs:   []*codectypes.Any{msgAny, failingMsgAny},
 			},