+++ types/query/filtered_pagination.go
@@ -3,7 +3,7 @@ package query
 import (
 	"fmt"
 
-	proto "github.com/cosmos/gogoproto/proto"
+	"github.com/cosmos/gogoproto/proto"
 
 	"cosmossdk.io/store/types"