+++ x/gov/keeper/vote.go
@@ -35,6 +35,11 @@ func (k Keeper) AddVote(ctx context.Context, proposalID uint64, voterAddr sdk.Ac
 		return err
 	}
 
+	err = k.assertVoteOptionsLen(options)
+	if err != nil {
+		return err
+	}
+
 	for _, option := range options {
 		switch proposal.ProposalType {
 		case v1.ProposalType_PROPOSAL_TYPE_OPTIMISTIC:
@@ -83,7 +88,7 @@ func (k Keeper) AddVote(ctx context.Context, proposalID uint64, voterAddr sdk.Ac
 		return err
 	}
 
-	return k.environment.EventService.EventManager(ctx).EmitKV(types.EventTypeProposalVote,
+	return k.EventService.EventManager(ctx).EmitKV(types.EventTypeProposalVote,
 		event.NewAttribute(types.AttributeKeyVoter, voterStrAddr),
 		event.NewAttribute(types.AttributeKeyOption, options.String()),
 		event.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposalID)),